Permissão de acesso

+++ app/install.sh
@@ -5,6 +5,11 @@ if [ "$EUID" -ne 0 ]; then
     exit
 fi
 
+if [ -L ${DIR} ] ; then
+    echo "O diretório ${DIR} nao pode ser usado pois é um link simbólico."
+    exit 0
+fi
+
 apt update -y
 apt install python3-pip -y
 apt install libxcb-* -y
@@ -19,5 +24,17 @@ pip3 install requests
 pip3 install PySide6
 pip3 install pycryptodome
 pip3 install colorama
+
+chmod +x /opt/codoencrypt/app
+chmod +x /opt/codoencrypt/app/gui
+chmod +x /opt/codoencrypt/app/gui/view
+chmod +r /opt/codoencrypt/app
+chmod +r /opt/codoencrypt/app/gui
+chmod +r /opt/codoencrypt/app/gui/view
+chmod +r /opt/codoencrypt/api
+chmod +x /opt/codoencrypt/api
+chmod +r /opt/codoencrypt/app/gui/codog
+chmod +x /opt/codoencrypt/app/gui/codog
+chmod +r /opt/codoencrypt/app/gui/codog.py
 
 ln -s /opt/codoencrypt/app/gui/codog /bin/codog
